+++ src/gui/gtk3/mod.rs
@@ -1334,7 +1334,7 @@ fn ui_init(app: &gtk::Application) {
                     }
                     GuiMessage::UpdateSensorValues(results) => {
                         debug!("Update sensor values with: {:?}", &results);
-                        gui.update_sensor_values(results);
+                        // gui.update_rreg_sensor_values(results);
                     }
                     GuiMessage::UpdateSerialPorts(ports) => {
                         debug!("Update Serial Ports with: {:?}", &ports);
@@ -1342,11 +1342,13 @@ fn ui_init(app: &gtk::Application) {
                     }
                     GuiMessage::UpdateRregs(results) => {
                         debug!("Update Rregs with: {:?}", &results);
-                        gui.update_rreg_store(results);
+                        gui.update_rreg_sensor_values(&results);
+                        gui.update_rreg_store(&results);
                     }
                     GuiMessage::UpdateRwregs(results) => {
                         debug!("Update Rwregs with: {:?}", &results);
-                        gui.update_rwreg_store(results);
+                        gui.update_rwreg_sensor_values(&results);
+                        gui.update_rwreg_store(&results);
                     }
                 }
             }
@@ -1565,57 +1567,58 @@ impl Gui {
         self.revealer_infobar_question.set_reveal_child(true);
     }
 
-    /// Update SensorValues
-    fn update_sensor_values(&self, result: Vec<(u16, u16)>) {
+    /// Update SensorValues mit den Werten der Lese-Register
+    fn update_rreg_sensor_values(&self, result: &[(u16, u16)]) {
         if let Ok(platine) = self.platine.lock() {
             if let Some(platine) = &*platine {
                 match platine.name() {
                 "Sensor-MB-CO2_O2_REV1_0" => {
-                    // Messzelle 1
+                    // Update Konzentration Messzelle 1
                     if let Some((_, value)) = result.get(2) {
                         self.label_sensor1_value_value.set_text(&value.to_string());
                     }
-                    // Messzelle 2
+                    // Update Konzentration Messzelle 2
                     if let Some((_, value)) = result.get(6) {
                         self.label_sensor2_value_value.set_text(&value.to_string());
                     }
                 },
                 "Sensor-MB-NAP5x_REV1_0" => {
-                    // Messzelle 1
+                    // Update Konzentration Messzelle 1
                     if let Some((_, value)) = result.get(2) {
                         self.label_sensor_value_value.set_text(&value.to_string());
                     }
                 },
                 "Sensor-MB-NAP5xx_REV1_0" => {
-                    // Messzelle 1
+                    // Update Konzentration Messzelle 1
                     if let Some((_, value)) = result.get(2) {
                         self.label_sensor1_value_value.set_text(&value.to_string());
                     }
-                    // Messzelle 2
+                    // Update Konzentration Messzelle 2
                     if let Some((_, value)) = result.get(6) {
                         self.label_sensor2_value_value.set_text(&value.to_string());
                     }
                 },
                 "Sensor-MB-NE4_REV1_0" => {
-                    // Messzelle 1
+                    // Update Konzentration Messzelle 1
                     if let Some((_, value)) = result.get(2) {
                         self.label_sensor_value_value.set_text(&value.to_string());
                     }
                 },
                 "Sensor-MB-NE4-V1.0" => {
-                    // Messzelle 1
+                    // Update Konzentration Messzelle 1
                     if let Some((_, value)) = result.get(2) {
                         self.label_sensor_value_value.set_text(&value.to_string());
                     }
                 },
                 "Sensor-MB-SP42A_REV1_0" => {
-                    // Messzelle 1
+                    // Update Konzentration Messzelle 1
                     if let Some((_, value)) = result.get(2) {
                         self.label_sensor_value_value.set_text(&value.to_string());
                     }
                 },
                 _ => self.show_infobar_error("Nicht unterstützte Platine, Sensorwerte konnten nicht aktualisiert werden."),
                 };
+                // Update Arbeitsweise
                 if let Some((_, value)) = result.get(1) {
                     self.combo_box_text_sensor_working_mode.set_active_id(Some(&format!("{}", value)));
                 }
@@ -1623,17 +1626,70 @@ impl Gui {
         }
     }
 
+        /// Update SensorValues mit den Werten der Lese-Register
+    fn update_rwreg_sensor_values(&self, result: &[(u16, u16)]) {
+        if let Ok(platine) = self.platine.lock() {
+            if let Some(platine) = &*platine {
+                match platine.name() {
+                "Sensor-MB-CO2_O2_REV1_0" => {
+                    // Update Modbus Adresse
+                    if let Some((_reg, address)) = result.iter().find(|&(reg, _address)| *reg == 80) {
+                        let modbus_address = self.spin_button_new_modbus_address.get_adjustment();
+                        modbus_address.set_value((*address).into());
+                    }
+                },
+                "Sensor-MB-NAP5x_REV1_0" => {
+                    // // Update Modbus Adresse
+                    if let Some((_reg, address)) = result.iter().find(|&(reg, _address)| *reg == 80) {
+                        let modbus_address = self.spin_button_new_modbus_address.get_adjustment();
+                        modbus_address.set_value((*address).into());
+                    }
+                },
+                "Sensor-MB-NAP5xx_REV1_0" => {
+                    // // Update Modbus Adresse
+                    if let Some((_reg, address)) = result.iter().find(|&(reg, _address)| *reg == 80) {
+                        let modbus_address = self.spin_button_new_modbus_address.get_adjustment();
+                        modbus_address.set_value((*address).into());
+                    }
+                },
+                "Sensor-MB-NE4_REV1_0" => {
+                    // // Update Modbus Adresse
+                    if let Some((_reg, address)) = result.iter().find(|&(reg, _address)| *reg == 80) {
+                        let modbus_address = self.spin_button_new_modbus_address.get_adjustment();
+                        modbus_address.set_value((*address).into());
+                    }
+                },
+                "Sensor-MB-NE4-V1.0" => {
+                    // // Update Modbus Adresse
+                    if let Some((_reg, address)) = result.iter().find(|&(reg, _address)| *reg == 50) {
+                        let modbus_address = self.spin_button_new_modbus_address.get_adjustment();
+                        modbus_address.set_value((*address).into());
+                    }
+                },
+                "Sensor-MB-SP42A_REV1_0" => {
+                    // // Update Modbus Adresse
+                    if let Some((_reg, address)) = result.iter().find(|&(reg, _address)| *reg == 80) {
+                        let modbus_address = self.spin_button_new_modbus_address.get_adjustment();
+                        modbus_address.set_value((*address).into());
+                    }
+                },
+                _ => self.show_infobar_error("Nicht unterstützte Platine, Sensorwerte konnten nicht aktualisiert werden."),
+                };
+            }
+        }
+    }
+
     /// Update RregStore
-    fn update_rreg_store(&self, result: Vec<(u16, u16)>) {
+    fn update_rreg_store(&self, result: &[(u16, u16)]) {
         if let Ok(lock) = self.rreg_store.lock() {
-            if let Some(ref store) = *lock { store.update_treestore(result) }
+            if let Some(ref store) = *lock { store.update_treestore(&result) }
         }
     }
 
     /// Update RwregStore
-    fn update_rwreg_store(&self, result: Vec<(u16, u16)>) {
+    fn update_rwreg_store(&self, result: &[(u16, u16)]) {
         if let Ok(lock) = self.rwreg_store.lock() {
-            if let Some(ref store) = *lock { store.update_treestore(result) }
+            if let Some(ref store) = *lock { store.update_treestore(&result) }
         }
     }
 
@@ -1783,7 +1839,7 @@ pub fn show_error(tx: &mpsc::Sender<GuiMessage>, msg: &str) {
 /// deren Funktionen wie `Gui::show_infobar_info` in den Callbacks nicht
 /// aufrufbar sind.
 /// Diese Funktion sended über den gui_tx Channel eine Nachricht an die InfoBar.
-pub fn _show_question(tx: &mpsc::Sender<GuiMessage>, msg: &str) {
+pub fn show_question(tx: &mpsc::Sender<GuiMessage>, msg: &str) {
     let timestamp = chrono::Local::now().format("%Y-%m-%d %H:%M:%S").to_string();
     tx.clone()
         .try_send(GuiMessage::ShowQuestion(format!(